Feature/qasm reader in openqldclab

+++ src/scheduler.h
@@ -120,7 +120,7 @@ public:
         ListDigraph::Node srcNode = graph.nodeFromId(srcID);
         ListDigraph::Node tgtNode = graph.nodeFromId(tgtID);
         ListDigraph::Arc arc = graph.addArc(srcNode,tgtNode);
-        weight[arc] = std::ceil( static_cast<float>(instruction[srcNode]->duration) / cycle_time);
+        weight[arc] = int(std::ceil( static_cast<float>(instruction[srcNode]->duration) / cycle_time));
         cause[arc] = operand;
         depType[arc] = deptype;
         DOUT("... dep " << name[srcNode] << " -> " << name[tgtNode] << " (opnd=" << operand << ", dep=" << DepTypesNames[deptype] << ")");
@@ -1051,8 +1051,8 @@ private:
                 if(platform.instruction_settings[id].count("latency") > 0)
                 {
                     float latency_ns = platform.instruction_settings[id]["latency"];
-                    latency_cycles = (std::ceil( static_cast<float>(std::abs(latency_ns)) / cycle_time)) *
-                                            ql::utils::sign_of(latency_ns);
+                    latency_cycles = long(std::ceil( static_cast<float>(std::abs(latency_ns)) / cycle_time)) *
+                                          ql::utils::sign_of(latency_ns);
                 }
             }
             cycle[*it] = cycle[*it] + latency_cycles;
@@ -1143,7 +1143,7 @@ private:
                     size_t iduration = ins->duration;
                     bduration = std::max(bduration, iduration);
                 }
-                abundle.duration_in_cycles = std::ceil(static_cast<float>(bduration)/cycle_time);
+                abundle.duration_in_cycles = size_t(std::ceil(static_cast<float>(bduration)/cycle_time));
                 bundles.push_back(abundle);
             }
         }
@@ -1219,7 +1219,7 @@ private:
                     bduration = std::max(bduration, iduration);
                 }
                 abundle.start_cycle = currCycle;
-                abundle.duration_in_cycles = std::ceil(static_cast<float>(bduration)/cycle_time);
+                abundle.duration_in_cycles = size_t(std::ceil(static_cast<float>(bduration)/cycle_time));
                 bundles.push_back(abundle);
             }
         }
@@ -1351,7 +1351,7 @@ private:
                 std::string operation_name(id);
                 std::string operation_type; // MW/FLUX/READOUT etc
                 std::string instruction_type; // single / two qubit
-                size_t operation_duration = std::ceil( static_cast<float>(curr_ins->duration) / cycle_time);
+                size_t operation_duration = size_t(std::ceil( static_cast<float>(curr_ins->duration) / cycle_time));
 
                 // FIXME: platform dependent fields
                 if(platform.instruction_settings.count(id) > 0)
@@ -1415,8 +1415,8 @@ private:
                 if(platform.instruction_settings[id].count("latency") > 0)
                 {
                     float latency_ns = platform.instruction_settings[id]["latency"];
-                    latency_cycles = (std::ceil( static_cast<float>(std::abs(latency_ns)) / cycle_time)) *
-                                            ql::utils::sign_of(latency_ns);
+                    latency_cycles = long(std::ceil( static_cast<float>(std::abs(latency_ns)) / cycle_time)) *
+                                          ql::utils::sign_of(latency_ns);
                 }
             }
             cycle[*it] = cycle[*it] + latency_cycles;
@@ -1605,7 +1605,7 @@ private:
                     size_t iduration = ins->duration;
                     bduration = std::max(bduration, iduration);
                 }
-                abundle.duration_in_cycles = std::ceil(static_cast<float>(bduration)/cycle_time);
+                abundle.duration_in_cycles = size_t(std::ceil(static_cast<float>(bduration)/cycle_time));
                 bundles.push_back(abundle);
             }
         }
@@ -1678,7 +1678,7 @@ private:
                     size_t iduration = ins->duration;
                     bduration = std::max(bduration, iduration);
                 }
-                abundle.duration_in_cycles = std::ceil(static_cast<float>(bduration)/cycle_time);
+                abundle.duration_in_cycles = size_t(std::ceil(static_cast<float>(bduration)/cycle_time));
                 bundles.push_back(abundle);
             }
         }
@@ -2036,7 +2036,7 @@ private:
                     size_t iduration = ins->duration;
                     bduration = std::max(bduration, iduration);
                 }
-                abundle.duration_in_cycles = std::ceil(static_cast<float>(bduration)/cycle_time);
+                abundle.duration_in_cycles = size_t(std::ceil(static_cast<float>(bduration)/cycle_time));
                 bundles.push_back(abundle);
             }
         }
@@ -2320,8 +2320,8 @@ private:
                 if(platform.instruction_settings[id].count("latency") > 0)
                 {
                     float latency_ns = platform.instruction_settings[id]["latency"];
-                    latency_cycles = (std::ceil( static_cast<float>(std::abs(latency_ns)) / cycle_time)) *
-                                            ql::utils::sign_of(latency_ns);
+                    latency_cycles = long(std::ceil( static_cast<float>(std::abs(latency_ns)) / cycle_time)) *
+                                          ql::utils::sign_of(latency_ns);
                     compensated_one = true;
 
                     gp->cycle = gp->cycle + latency_cycles;
@@ -2815,7 +2815,7 @@ private:
             std::string operation_name;
             std::string operation_type;
             std::string instruction_type;
-            size_t      operation_duration = std::ceil( static_cast<float>(gp->duration) / cycle_time);
+            size_t      operation_duration = size_t(std::ceil( static_cast<float>(gp->duration) / cycle_time));
             GetGateParameters(gp->name, platform, operation_name, operation_type, instruction_type);
             if (rm.available(curr_cycle, gp, operation_name, operation_type, instruction_type, operation_duration))
             {
@@ -2928,7 +2928,7 @@ private:
                 size_t      operation_duration = 0;
 
                 GetGateParameters(gp->name, platform, operation_name, operation_type, instruction_type);
-                operation_duration = std::ceil( static_cast<float>(gp->duration) / cycle_time);
+                operation_duration = size_t(std::ceil( static_cast<float>(gp->duration) / cycle_time));
                 rm.reserve(curr_cycle, gp, operation_name, operation_type, instruction_type, operation_duration);
             }
             TakeAvailable(selected_node, avlist, scheduled, dir);   // update avlist/scheduled/cycle
@@ -3107,7 +3107,7 @@ private:
                     DOUT("... considering: " << predgp->qasm() << " @cycle=" << predgp->cycle << " remaining=" << remaining[pred_node]);
 
                     // candidate's result, when moved, must be ready before end-of-circuit and before used
-                    predgp_completion_cycle = curr_cycle + std::ceil(static_cast<float>(predgp->duration)/cycle_time);
+                    predgp_completion_cycle = curr_cycle + size_t(std::ceil(static_cast<float>(predgp->duration)/cycle_time));
                     if (predgp_completion_cycle > cycle_count + 1)  // at SINK is ok, later not
                     {
                         forward_predgp = false;
